requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-paypal==2.1
djangorestframework==3.15.2
idna==3.10
mysqlclient==2.2.7
paypal==1.2.5
pillow==11.1.0
pytz==2025.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
